Instructions for Syncing Your Changes (READMEpls)
After making changes to the ButtonMashStudios folder, open GitHub
desktop and click "ButtonMashStudios" in the left Dashboard.

Changes should automatically be populated. Click the text boxes on the
bottom of the window. Give a name and summary of what your changes were.
You won't be able to click "Commit to master" without a title and
description.

Click Commit to Master and then press Sync in the top right of the
window for good measure.